feat: Set fake data with css and Apply React draggable
When we work together to make Front-end and Back-end server,
as front-end developer, you should be prepared for real data.
To test if the datas can be displayed properly, I used @faker-js/faker.
And i saw the design doesn't look that satisfied for me. So i changed
it.

I want to move Model component like drag and drop functionality.
I found out React draggable could be a good option for this.
I'm in the middle of applying this library, stil needed to understand.