feat: Add modal drag and Utilize modal funtionality.
I found out that 'react-draggable' can help us to make this component.
In order to share this and reuse it in other components, i seperated
this into utilComponent in two part.
DraggableModal.jsx in utilComponent folder make blueprint of modal.
DraggableContent.jsx in the same folder make children of
DraggableModal.jsx
Of course, you can make another content if you're not satisfied with
DraggableModalContent.jsx, which is now used as children of
topcomponent. Please see through my code!

- pros and cons
  pros : i can share this component (reusability goes up)
  cons : little bit hard to understand i guess??

I think the important part of making util component is checking the
types of properties. For this, i also applied PropTypes.

And back-end codes are changed a little bit.(annotations)

Resolve: #1