feat: Make test code for MemberChattingRoom.java
I was quite wondering how to use MemberChattingRoom.java as a centre of Member.java and ChattingRoom.java. So, I made test code(need more modification).

In the most company, they don't delete data right away. They use soft-delete at first, and it depends on whether they really remove it or not. I think it's better to apply this to all Entity, then I changed BaseEntity code for this.