feat: Use JPA and queryDsl together.
I used to use this.But I almost forgot how to do this.
If you want to use JPA functionalities (basic crud things) and
add customized QueryDSL codes, follow this step.

1. make interface that extends JpaRepository, which is used for JPA.
2. make ~~Impl class extends SimpleJpaRepository and implement the interface the above.
3. If you import ~~Impl file and code assist, you can use JPA default functionalities and customized QueryDSL's method.

it's not that difficult. But I forget it. So I write this down here and share this.

example : MemberChattingRoomRepositoryImpl