feat: Make refresh functionality with RefreshToken
I seem to have overlooked one import thing for JWT.
It's refresh functionality. I'm in the middle of making this one and I'll be soon done by the end of the day today. You can get the process of how JWT must do when user login in special case.

And I made special enum type(ContentType) and JsonUtil.

See also: #13