feat: Add Message entity and comments
I get to the conclusion that it's not a perfect choice to use spring websocket for chatting app.
So, after I make enough test with it, I'll introduce the another message broker.
But before that, We need to make sure that it works well regardless of what message broker we use. For this, I made new Message entity.

I realized how web socket works. While studying how it works,
I wrote some comments of how each method work in web socket codes.

Also, there are minor modifications for propTypes and removing unused codes.