feat: Design chatting room and apply Lotout func.
I hope following the whole design of Kakao talk as much as possible.
So, I applied new design.
Unfortunately, there was an mistake.
When I pushed codes, I guess It overwrote codes. I didn' mean it but it
happened So I tried to remake proper codes.

I'm on my way to make logout func and apply refresh func to front-end
with communication of back-end server.