feat: Add loginCheck api but test needed
The reason we use JWT is not to give too much burdens to server just for authentication.
But, in order to keep our jwt safe in cookies, if we serve tokens in cookies, HTTP-Only cookies are recommended.
Problem of this is that we can't get the cookie in javascript. So, we must check if there's cookie that we can judge he already login by cookie.