feat: Add redirection func in React and reissue access token (modified)
I've been dealing with redirection when server side response is not proper. Before I lead user to another page, I need to check many cases when this really happens.
I encountered a lot of problems

  1. I removed default redirect lead when error occur to '/error' page, which doesn't used for me. It causes additional filtering work after I check the whole tokens because I also throw the error when I check refresh token again.

  2. I needed to add another action in redux-saga for guidance of proper url when error occur. I wanted to guide user to go to the 'signin' page when access token and refresh token are expired(If we don't have any tools to authenticate user, I think it's reasonable!).

  3. I was wondering tons of times because of this question.
 "Do I really need to make all the cookie to have HTTPOnly option?"
 This question make me feel irritated because as it turned out, it just depends on situation.
 If we're not satisfied with security and really worried to issue your cookie much safer(attack like XSS), It's good choice to set all cookies as HTTPOnly. But It's noteworthy to know this might lead more complexity for your code. Access token's expired date is so short that it's quite safe. Even if hackers still cookie, they don't know key to decode it!
In my case, I tried to apply only HTTPOnly cookies for practice.
But still don't know exactly what is the best for me.
Just I focused on security more and gave more burdens to sever side.

It was a long journey for me. My boss asked me to do too much works, which couldn't be done without overworks. And I already made too much appointment with my coworkers. Anyway, whether it's late or not, or it's too late or not, important thing is that i finally made it.

Resolve: #19, #22