feat: Add error alert when log in and new Component added
I found out something is going wrong when i log in.
Of course, I didn't sign up for that error but user who want to log in have two refresh token. I don't allow the user to have two refresh token.
Anyway, when error occur in server side, we need to show what error occur and try it again or just ask to company things like that. I fixed it!

I also added new draggable modal for new multi profile of user.
Of course the design of it is quite similar with previous one.
But the objective of that one is a little bit different. First one is for just showing information, but the new one is for adding new profile.
I wanted to separate component based on objective.