style: Design common modal and add buttons for ops
I've been trying to imitate kakao talk applcation.
I dicided to use CommonProfileModalContent, shared with many components,
for a variety of cases.
For instance, When I open multi profile, design of it is same with the
one that I open in fried item. That's because the whole design of modal
is not that different except for some icons.
So, It's quite boring to make another one even if i can just reuse it
with flag.