feat: Add birthday display style and modal
I finally make it up!
When user want to know who are birthday on this month, he can know it by
clicking birthday cake icon.

Great thing of this component is that grouping with reduce in JS is
amazing. I was not that curious and even don't know how to handle these
kinds of default functions.
I also made a lot of DateUtil functions. Just check it out.