style: modify profile css
I changed class name to share the same styles when user see user's
profile.

I'll apply this to FriendItem.jsx file sometime soon.

And I think I need to think of checking who updated his profile.
In order to check this, we should save data for this somewhere.
I think I've almost forgot of it.
So, I need to discuess it with GPT.